ui and interceptor fix

+++ app/src/main/java/com/berkedursunoglu/week4assignmentbd/model/Minutely.kt
@@ -2,5 +2,5 @@ package com.berkedursunoglu.week4assignmentbd.model
 
 data class Minutely(
     val dt: Int,
-    val precipitation: Int
+    val precipitation: Double
 )
